requirements.txt
attrs==21.2.0
aws-cdk.assets==1.118.0
aws-cdk.aws-apigateway==1.118.0
aws-cdk.aws-applicationautoscaling==1.118.0
aws-cdk.aws-autoscaling==1.118.0
aws-cdk.aws-autoscaling-common==1.118.0
aws-cdk.aws-autoscaling-hooktargets==1.118.0
aws-cdk.aws-certificatemanager==1.118.0
aws-cdk.aws-cloudformation==1.118.0
aws-cdk.aws-cloudfront==1.118.0
aws-cdk.aws-cloudwatch==1.118.0
aws-cdk.aws-codebuild==1.118.0
aws-cdk.aws-codecommit==1.118.0
aws-cdk.aws-codeguruprofiler==1.118.0
aws-cdk.aws-codepipeline==1.118.0
aws-cdk.aws-codestarnotifications==1.118.0
aws-cdk.aws-cognito==1.118.0
aws-cdk.aws-ec2==1.118.0
aws-cdk.aws-ecr==1.118.0
aws-cdk.aws-ecr-assets==1.118.0
aws-cdk.aws-ecs==1.118.0
aws-cdk.aws-ecs-patterns==1.118.0
aws-cdk.aws-efs==1.118.0
aws-cdk.aws-elasticloadbalancing==1.118.0
aws-cdk.aws-elasticloadbalancingv2==1.118.0
aws-cdk.aws-events==1.118.0
aws-cdk.aws-events-targets==1.118.0
aws-cdk.aws-globalaccelerator==1.118.0
aws-cdk.aws-iam==1.118.0
aws-cdk.aws-kinesis==1.118.0
aws-cdk.aws-kinesisfirehose==1.118.0
aws-cdk.aws-kms==1.118.0
aws-cdk.aws-lambda==1.118.0
aws-cdk.aws-logs==1.118.0
aws-cdk.aws-rds==1.118.0
aws-cdk.aws-route53==1.118.0
aws-cdk.aws-route53-targets==1.118.0
aws-cdk.aws-s3==1.118.0
aws-cdk.aws-s3-assets==1.118.0
aws-cdk.aws-sam==1.118.0
aws-cdk.aws-secretsmanager==1.118.0
aws-cdk.aws-servicediscovery==1.118.0
aws-cdk.aws-signer==1.118.0
aws-cdk.aws-sns==1.118.0
aws-cdk.aws-sns-subscriptions==1.118.0
aws-cdk.aws-sqs==1.118.0
aws-cdk.aws-ssm==1.118.0
aws-cdk.aws-stepfunctions==1.118.0
aws-cdk.cloud-assembly-schema==1.118.0
aws-cdk.core==1.118.0
aws-cdk.custom-resources==1.118.0
aws-cdk.cx-api==1.118.0
aws-cdk.region-info==1.118.0
backports.entry-points-selectable==1.1.0
cattrs==1.7.1
constructs==3.3.124
distlib==0.3.2
filelock==3.0.12
jsii==1.33.0
platformdirs==2.2.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
-e git+ssh://git@github.com/TASUKI-Corporation/touch-and-plan-cdk.git@6ce06623965b7639fc5536fd133408d3c8e158f3#egg=touch_and_plan_cdk
typing-extensions==3.10.0.0
virtualenv==20.7.2
